linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -14,7 +14,7 @@ int linear_search(int *array, size_t size, int value)
 
 	if (array == NULL || size == 0)
 		return (-1);
-	while (itr < size)
+	while (itr < (int)size)
 	{
 		printf("Value checked array[%d] = [%d]\n", itr, array[itr]);
 		if (array[itr] == value)
